PHpstan and PHpDoc

+++ src/Controller/LibraryController.php
@@ -8,16 +8,17 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
-
 use App\Entity\Book;
 use App\Repository\BookRepository;
+use SebastianBergmann\Environment\Console;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class LibraryController extends AbstractController
 {
     #[Route("/library", name: "library_index")]
     public function index(): Response
-    {   
+    {
         return $this->render("library/index.html.twig");
     }
 
@@ -37,7 +38,7 @@ class LibraryController extends AbstractController
 
     #[Route("/library/book/{num}", name: "read_one", requirements: ['num' => '\d+'])]
     public function viewBook(BookRepository $bookRepository, int $num): Response
-    {   
+    {
         $book = $bookRepository->find($num);
 
         $data = [
@@ -50,7 +51,7 @@ class LibraryController extends AbstractController
 
     #[Route("/library/book/delete/{num}", name: "confirm_delete")]
     public function confirmDelete(BookRepository $bookRepository, int $num, Request $request): Response
-    {   
+    {
         $book = $bookRepository->find($num);
 
         $data = [
@@ -63,12 +64,16 @@ class LibraryController extends AbstractController
 
     #[Route("/library/book/delete", name: "delete_book")]
     public function deleteBook(BookRepository $bookRepository, Request $request, EntityManagerInterface $entityManager): Response
-    {   
+    {
         $bookId = $request->request->get("book_id");
         if (!$bookId) {
             return $this->redirectToRoute("index");
         }
         $book = $bookRepository->find($bookId);
+
+        if (!$book) {
+            throw new NotFoundResourceException("Hittar ingen bok med id $bookId");
+        }
 
         $entityManager->remove($book);
 
@@ -83,7 +88,7 @@ class LibraryController extends AbstractController
 
     #[Route("/library/book/update/{num}", name: "init_update_book", requirements: ['num' => '\d+'], methods:["GET"])]
     public function initUpdateBook(BookRepository $bookRepository, int $num): Response
-    {   
+    {
         $book = $bookRepository->find($num);
 
         $data = [
@@ -96,25 +101,35 @@ class LibraryController extends AbstractController
 
     #[Route("/library/book/update", name: "update_book", methods:["POST"])]
     public function updateBook(BookRepository $bookRepository, EntityManagerInterface $entityManager, Request $request): Response
-    {   
+    {
         $bookId = $request->request->get("book_id");
-        
+
         $book = $bookRepository->find($bookId);
 
         if (!$book) {
             throw new NotFoundHttpException("Book with id $bookId cannot be found!");
         }
 
-        // Uppdatera objekt
-        $book->setTitle($request->request->get("title"));
-        $book->setIsbn($request->request->get("isbn"));
-        $book->setAuthor($request->request->get("author"));
-        $book->setPublisher($request->request->get("publisher"));
-        $imgUrl = $request->request->get("img_url");
-        $imgUrlFull = $imgUrl ? $imgUrl : "na.png";
-        $book->setImgUrl($imgUrlFull);
+        if (array_all($_POST, function ($request) {
+            return is_string($request);
 
-        // Uppdatera databas
+        })) {
+            $title = (string)($_POST["title"]);
+            $isbn =   (string)($_POST["isbn"]);
+            $author = (string)($_POST["author"]);
+            $publisher = (string)($_POST["publisher"]);
+
+            // Uppdatera objekt, fält är required i form.
+            $book->setTitle($title);
+            $book->setIsbn($isbn);
+            $book->setAuthor($author);
+            $book->setPublisher($publisher);
+            $imgUrl = $_POST["img_url"];
+            $imgUrlFull = $imgUrl ? $imgUrl : "na.png";
+            $book->setImgUrl($imgUrlFull);
+        };
+
+        // // Uppdatera databas
         $entityManager->persist($book);
         $entityManager->flush();
 
@@ -126,7 +141,7 @@ class LibraryController extends AbstractController
 
     #[Route("/library/book/register", name: "register_book", methods:["GET"])]
     public function registerBook(): Response
-    {   
+    {
         return $this->render("library/register-book.html.twig");
     }
 
@@ -135,21 +150,33 @@ class LibraryController extends AbstractController
      * Add new book to the database
      */
     #[Route("/library/book/add", name: "add_book", methods:["POST"])]
-    public function addBook(Request $request, EntityManagerInterface $entityManager): Response
-    {   
-        $request->request->get("title");
-        $book = new Book();
-        $book->setTitle($request->request->get("title"));
-        $book->setIsbn($request->request->get("isbn"));
-        $book->setAuthor($request->request->get("author"));
-        $book->setPublisher($request->request->get("publisher"));
-        
-        $img = $request->request->get("img_url");
-        $book->setImgUrl($img ? $img . ".png" : "na.png");
-        
-        $entityManager->persist($book);
+    public function addBook(EntityManagerInterface $entityManager): Response
+    {
+        if (array_all($_POST, function ($request) {
+            return is_string($request);
+        })) {
+            $title = (string)($_POST["title"]);
+            $isbn =   (string)($_POST["isbn"]);
+            $author = (string)($_POST["author"]);
+            $publisher = (string)($_POST["publisher"]);
 
-        $entityManager->flush();
+            // Skapa Book objekt, fält är required i form.
+            $book = new Book();
+            $book->setTitle($title);
+            $book->setIsbn($isbn);
+            $book->setAuthor($author);
+            $book->setPublisher($publisher);
+            $imgUrl = $_POST["img_url"];
+            $imgUrlFull = $imgUrl ? $imgUrl : "na.png";
+            $book->setImgUrl($imgUrlFull);
+
+            $entityManager->persist($book);
+
+            $entityManager->flush();
+
+            $this->addFlash('notice', "Added book: $title");
+        }
         return $this->redirectToRoute("book_view_all");
+
     }
 }
